refactor(calendar): move calendar types and CalendarTimetableSyncDialog fix: wild apostrophe
